images/simulator/requirements.txt
python-can
cantools